Major changes
Think this post can be a guideline with certain links.

See the first bit for how I imagine it.

Do you think this is the way to go?

Need to find good external examples. Also.. text should be rewritten and
more internal examples should be given for each header.

This will also serve as an editorial guideline for us so it's important
to put some effort into it.