4th sentence & what vs which
You almost answered your own question when you wrote: "What is the best
way to explain.?" What is used to ask a question when there are an
unknown number or infinite possibilities for an answer. You know that
there are many, many ways that exist to address your question, and you
want to find out—from all those possibilities that you might not even
know about—what the best way is.

You can use which when you have a very small or limited field to choose
from. Certainly use which, not what, when there are only two choices, or
if both speaker and listener can visualize all the items under
consideration:

(c)

Which foot did you break?

(d)

Which one of your sisters is the singer?