created twitter message with image
created a twitter message with a image, the url will change depending
on the blogpost url. I did not make bit.ly because i think it is
important people see the url if they don’t klick the link. The goal of
the message is to spark enough curiosity without giving away too much
what we are doing.