added a image of the industrial revolution
the idea behind the image is that we now again face a sort of
industrial revolution. If i would add a image of a albertheijn, it
feels like it is too much about the alberthein. That is why i went fort
this. It would be interesting to treat this article in comparison to
education during the industrial revolution!