Added potential image
My thought process towards this image was:
-Read the pitch and article
-themes that stood out were growth, development, evolution
-Started hunting for images related to these themes—what I found were
unsatisfying or cliché images of human evolution silhouettes, charts
that were beautiful but too far from the original idea to make an
adequate connection, cheezy progressions of sprouting and growing
plants, terrible graphics trying to illustrate development, light
bulbs.
-I decided to merge two of the unsatisfactory results in hopes of
finding something better—evolution and light bulbs—and started looking
for 'evolution of the light bulb' and 'history of the light bulb' images

I tried to illustrate this in the commits.