i wrote a introduction
goal of achievable: To be amazed by the diversity of contexts in which
service design has an impact on human behavior. I wrote an example of
how drinking wine is influenced by the shape of the glass, even tho we
are often not aware of it. Wine can be replaced by any liquid