moved the file to the right folder and added a introduction
the goal of the introduction is to show the different context’s in
which service design plays a role. I chose a very simple and relatable
subject in order to make it super duper clear what we want. Maybe too
simple but hey! thats how i role. i tried to make a smooth transition
between the intro and the two scenarios