create new pitch and pitch format
So….

I tried to come up with a pitch and immediately tried to come up with
idea’s which is the end result, and thus the opposite of goal oriented
working. I took one outcome from the brief, expanded on it by writing
down exactly what it means, and then filtering the things i wanted to
work on. So, creating new audience has in theory this in it: Introduce
recurring new readers to the blog who will share the story and/or who
don't know [Client name] and/or [Core service/product] and could use
the [Core service/product] and/or work for/with [Client name]. And i
took out some things to make it less broad, this does not mean that we
will not achieve this things, but we are not actively looking for it. I
tried to make chapters that flowed natural from the goal. What do you
need the reader to experience in order to achieve the goal, and from
there how can we  doe this, why is this solution a good idea and what
assumptions is this idea based on