Attempted to add value by removing explicit metaphors.
I think the two scenarios are better introductory bait. The previous
introduction made the narrative arc into more of a narrative wavy line.

I took one paragraph of the introduction and moved it below the
scenarios as a better transition into the rest of the post, and to
explain the idea of 'context'.