Edit image
Because we may 😋 (LCC)
I cropped it because there was a distracting hand with camera in the
corner. Then I cropped it so that the head of the second window cleaner
was cut off to make it more mysterious (for the same goal I used a
filter and slightly adjust the lighting).