Image of a Caveman
The *Appeal* part talks about: *emotional reactions are some of the
strongest connections we can make to visual content. They occur from
the part of the brain that is also related to our survival instincts,
which means that the reaction may be more direct and intense.*

I thought of when we needed survival instincts, back in the
Neanderthaler days. First searched for gifs which illustrated this
visual survival instinct. Couldn’t find any suited.

Searched for cavemen, stumbled on this one on Flickr. First I didn’t
think it said enough about the content.

But now I do because the paragraph talks about: *Visual content should
indicate the emotion for which it aims. This could be accomplished
abstractly with colour, **metaphors**, or **irony**, for example. Or it
could be achieved concretely with associations which elicit the "right"
feeling (safety, **food**, positivity, love…).*