addd an image
i wanted something that has to do with robots and humans, and their relation. Stumbled on this image. Was quite funny, in movies robots often revolt vs humans, and we are pleading to create more of them. This is an airy way to speak to the imagination and create a context for the article that is memorable.